move timer to existing debug.py file
