+++ verl/workers/sharding_manager/megatron_sglang.py
@@ -27,8 +27,8 @@ from torch.distributed.device_mesh import DeviceMesh
 
 from verl.protocol import DataProto, all_gather_data_proto
 from verl.utils.debug import GPUMemoryLogger, log_gpu_memory_usage
+from verl.utils.debug.performance import _timer
 from verl.utils.megatron_utils import per_tensor_generator
-from verl.utils.timing_utils import _timer
 
 from .base import BaseShardingManager
