+++ verl/workers/sharding_manager/megatron_vllm.py
@@ -30,10 +30,10 @@ from verl.protocol import all_gather_data_proto
 from verl.third_party.vllm import LLM, vllm_version
 from verl.third_party.vllm import parallel_state as vllm_ps
 from verl.utils.debug import GPUMemoryLogger
+from verl.utils.debug.performance import _timer
 from verl.utils.megatron_utils import (
     per_tensor_generator,
 )
-from verl.utils.timing_utils import _timer
 from verl.utils.torch_functional import check_cuda_is_available
 from verl.utils.vllm_utils import patch_vllm_moe_model_weight_loader
